// src/app/features/builder/model/filed/i-field.ts
export class Field {
  _id?: any;
  name?: any;
  type?: any;
  icon?: any;
  toggle?: any;
  required?: any;
  regex?: any;
  errorText?: any;
  label?: any;
  sublabel?: any;
  description?: any;
  placeholder?: any;
  className?: any;
  subtype?: any;
  nestedType?: any;
  handle?: any;
  min?: number;
  max?: number;
  inline?: any;
  value?: any;
  values?: Array<value>;
  parent?: any;
  relatedTo?: any;
  category?: any;
  material?: any;
  hideRequiredMarker?: any;
  appearance?: any;
  hintLabel?: any;
  maxLength?: any;
  align?: any;
  size?: any;
  divider?: any;
  minDate?: any;
  maxDate?: any;
}
  export class value {
    label?: any = '';
    value?: any = '';
  }
  